phase-5/services/audit_service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
kafka-python>=2.0.2
dapr-ext-fastapi>=1.9.0
sqlmodel>=0.0.14
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.2